+++ backstage.json
@@ -1,3 +1,3 @@
 {
-  "version": "1.20.3"
+  "version": "1.21.1"
 }
