+++ package.json
@@ -31,7 +31,7 @@
     ]
   },
   "devDependencies": {
-    "@backstage/cli": "^0.24.0",
+    "@backstage/cli": "^0.25.0",
     "@spotify/prettier-config": "^12.0.0",
     "concurrently": "^8.0.0",
     "eslint-plugin-jest": "*",
