+++ packages/app/package.json
@@ -18,31 +18,31 @@
     "cy:run": "cypress run --browser chrome"
   },
   "dependencies": {
-    "@backstage/app-defaults": "^1.4.5",
+    "@backstage/app-defaults": "^1.4.6",
     "@backstage/catalog-model": "^1.4.3",
-    "@backstage/cli": "^0.24.0",
-    "@backstage/core-app-api": "^1.11.1",
-    "@backstage/core-components": "^0.13.8",
-    "@backstage/core-plugin-api": "^1.8.0",
-    "@backstage/integration-react": "^1.1.21",
-    "@backstage/plugin-api-docs": "^0.10.1",
-    "@backstage/plugin-catalog": "^1.15.1",
-    "@backstage/plugin-catalog-common": "^1.0.18",
-    "@backstage/plugin-catalog-graph": "^0.3.1",
-    "@backstage/plugin-catalog-import": "^0.10.3",
-    "@backstage/plugin-catalog-react": "^1.9.1",
-    "@backstage/plugin-github-actions": "^0.6.8",
-    "@backstage/plugin-org": "^0.6.17",
-    "@backstage/plugin-permission-react": "^0.4.17",
-    "@backstage/plugin-scaffolder": "^1.16.1",
-    "@backstage/plugin-search": "^1.4.3",
-    "@backstage/plugin-search-react": "^1.7.3",
-    "@backstage/plugin-tech-radar": "^0.6.10",
-    "@backstage/plugin-techdocs": "^1.9.1",
-    "@backstage/plugin-techdocs-module-addons-contrib": "^1.1.2",
-    "@backstage/plugin-techdocs-react": "^1.1.13",
-    "@backstage/plugin-user-settings": "^0.7.13",
-    "@backstage/theme": "^0.4.4",
+    "@backstage/cli": "^0.25.0",
+    "@backstage/core-app-api": "^1.11.2",
+    "@backstage/core-components": "^0.13.9",
+    "@backstage/core-plugin-api": "^1.8.1",
+    "@backstage/integration-react": "^1.1.22",
+    "@backstage/plugin-api-docs": "^0.10.2",
+    "@backstage/plugin-catalog": "^1.16.0",
+    "@backstage/plugin-catalog-common": "^1.0.19",
+    "@backstage/plugin-catalog-graph": "^0.3.2",
+    "@backstage/plugin-catalog-import": "^0.10.4",
+    "@backstage/plugin-catalog-react": "^1.9.2",
+    "@backstage/plugin-github-actions": "^0.6.9",
+    "@backstage/plugin-org": "^0.6.18",
+    "@backstage/plugin-permission-react": "^0.4.18",
+    "@backstage/plugin-scaffolder": "^1.17.0",
+    "@backstage/plugin-search": "^1.4.4",
+    "@backstage/plugin-search-react": "^1.7.4",
+    "@backstage/plugin-tech-radar": "^0.6.11",
+    "@backstage/plugin-techdocs": "^1.9.2",
+    "@backstage/plugin-techdocs-module-addons-contrib": "^1.1.3",
+    "@backstage/plugin-techdocs-react": "^1.1.14",
+    "@backstage/plugin-user-settings": "^0.7.14",
+    "@backstage/theme": "^0.5.0",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
     "@parsifal-m/plugin-dev-quotes-homepage": "^3.0.3",
@@ -54,7 +54,7 @@
     "react-use": "^17.2.4"
   },
   "devDependencies": {
-    "@backstage/test-utils": "^1.4.5",
+    "@backstage/test-utils": "^1.4.6",
     "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^14.0.0",
