+++ packages/backend/package.json
@@ -16,26 +16,28 @@
     "build-image": "docker build ../.. -f Dockerfile --tag backstage"
   },
   "dependencies": {
-    "@backstage/backend-common": "^0.19.9",
-    "@backstage/backend-tasks": "^0.5.12",
-    "@backstage/catalog-client": "^1.4.6",
+    "@backstage/backend-common": "^0.20.0",
+    "@backstage/backend-defaults": "^0.2.8",
+    "@backstage/backend-plugin-api": "^0.6.8",
+    "@backstage/backend-tasks": "^0.5.13",
+    "@backstage/catalog-client": "^1.5.1",
     "@backstage/catalog-model": "^1.4.3",
     "@backstage/config": "^1.1.1",
-    "@backstage/plugin-app-backend": "^0.3.55",
-    "@backstage/plugin-auth-backend": "^0.20.0",
-    "@backstage/plugin-auth-node": "^0.4.1",
-    "@backstage/plugin-catalog-backend": "^1.15.0",
-    "@backstage/plugin-catalog-backend-module-github": "^0.4.5",
+    "@backstage/plugin-app-backend": "^0.3.56",
+    "@backstage/plugin-auth-backend": "^0.20.2",
+    "@backstage/plugin-auth-node": "^0.4.2",
+    "@backstage/plugin-catalog-backend": "^1.16.0",
+    "@backstage/plugin-catalog-backend-module-github": "^0.4.6",
     "@backstage/plugin-catalog-common": "^1.0.19",
-    "@backstage/plugin-permission-backend": "^0.5.30",
-    "@backstage/plugin-permission-common": "^0.7.10",
-    "@backstage/plugin-permission-node": "^0.7.18",
-    "@backstage/plugin-proxy-backend": "^0.4.5",
-    "@backstage/plugin-scaffolder-backend": "^1.19.1",
-    "@backstage/plugin-search-backend": "^1.4.7",
-    "@backstage/plugin-search-backend-module-pg": "^0.5.16",
-    "@backstage/plugin-search-backend-node": "^1.2.11",
-    "@backstage/plugin-techdocs-backend": "^1.9.0",
+    "@backstage/plugin-permission-backend": "^0.5.31",
+    "@backstage/plugin-permission-common": "^0.7.11",
+    "@backstage/plugin-permission-node": "^0.7.19",
+    "@backstage/plugin-proxy-backend": "^0.4.6",
+    "@backstage/plugin-scaffolder-backend": "^1.19.2",
+    "@backstage/plugin-search-backend": "^1.4.8",
+    "@backstage/plugin-search-backend-module-pg": "^0.5.17",
+    "@backstage/plugin-search-backend-node": "^1.2.12",
+    "@backstage/plugin-techdocs-backend": "^1.9.1",
     "@parsifal-m/opa-permissions-wrapper": "^0.3.1",
     "@parsifal-m/plugin-opa-backend": "^1.1.3",
     "app": "link:../app",
@@ -47,7 +49,7 @@
     "winston": "^3.2.1"
   },
   "devDependencies": {
-    "@backstage/cli": "^0.24.0",
+    "@backstage/cli": "^0.25.0",
     "@types/dockerode": "^3.3.0",
     "@types/express": "^4.17.6",
     "@types/express-serve-static-core": "^4.17.5",
