+++ plugins/backstage-opa-backend/package.json
@@ -33,7 +33,8 @@
     "postpack": "backstage-cli package postpack"
   },
   "dependencies": {
-    "@backstage/backend-common": "^0.19.9",
+    "@backstage/backend-common": "^0.20.0",
+    "@backstage/backend-plugin-api": "^0.6.8",
     "@backstage/config": "^1.1.1",
     "@backstage/errors": "^1.2.3",
     "@types/express": "*",
@@ -44,7 +45,7 @@
     "yn": "^4.0.0"
   },
   "devDependencies": {
-    "@backstage/cli": "^0.24.0",
+    "@backstage/cli": "^0.25.0",
     "@types/supertest": "^2.0.12",
     "msw": "^1.0.0",
     "supertest": "^6.2.4"
