+++ plugins/backstage-opa-backend/src/index.ts
@@ -1,1 +1,2 @@
 export * from './service/router';
+export { opaBackendPlugin as default } from './plugin';
