+++ plugins/backstage-opa-entity-checker/package.json
@@ -36,10 +36,10 @@
   "dependencies": {
     "@backstage/catalog-model": "^1.4.3",
     "@backstage/config": "^1.1.1",
-    "@backstage/core-components": "^0.13.8",
-    "@backstage/core-plugin-api": "^1.8.0",
-    "@backstage/plugin-catalog-react": "^1.9.1",
-    "@backstage/theme": "^0.4.4",
+    "@backstage/core-components": "^0.13.9",
+    "@backstage/core-plugin-api": "^1.8.1",
+    "@backstage/plugin-catalog-react": "^1.9.2",
+    "@backstage/theme": "^0.5.0",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "4.0.0-alpha.61",
@@ -50,10 +50,10 @@
     "react": "^16.13.1 || ^18.0.0"
   },
   "devDependencies": {
-    "@backstage/cli": "^0.24.0",
-    "@backstage/core-app-api": "^1.11.1",
-    "@backstage/dev-utils": "^1.0.24",
-    "@backstage/test-utils": "^1.4.5",
+    "@backstage/cli": "^0.25.0",
+    "@backstage/core-app-api": "^1.11.2",
+    "@backstage/dev-utils": "^1.0.25",
+    "@backstage/test-utils": "^1.4.6",
     "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^14.0.0",
