+++ plugins/backstage-opa-permissions-wrapper/package.json
@@ -33,14 +33,14 @@
     "postpack": "backstage-cli package postpack"
   },
   "dependencies": {
-    "@backstage/backend-common": "^0.19.9",
+    "@backstage/backend-common": "^0.20.0",
     "@backstage/config": "^1.1.1",
     "@backstage/errors": "^1.2.3",
-    "@backstage/plugin-auth-node": "^0.4.1",
-    "@backstage/plugin-catalog-backend": "^1.15.0",
-    "@backstage/plugin-permission-common": "^0.7.10",
-    "@backstage/plugin-permission-node": "^0.7.18",
-    "@backstage/plugin-scaffolder-backend": "^1.19.1",
+    "@backstage/plugin-auth-node": "^0.4.2",
+    "@backstage/plugin-catalog-backend": "^1.16.0",
+    "@backstage/plugin-permission-common": "^0.7.11",
+    "@backstage/plugin-permission-node": "^0.7.19",
+    "@backstage/plugin-scaffolder-backend": "^1.19.2",
     "@types/express": "*",
     "express": "^4.17.3",
     "express-promise-router": "^4.1.0",
@@ -49,7 +49,7 @@
     "yn": "^4.0.0"
   },
   "devDependencies": {
-    "@backstage/cli": "^0.24.0",
+    "@backstage/cli": "^0.25.0",
     "@types/supertest": "^2.0.12",
     "msw": "^0.49.0",
     "supertest": "^6.2.4"
